requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cytoolz==0.11.2
dataclassy==0.11.1
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.7
eth-brownie==1.18.1
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
frozenlist==1.3.0
hexbytes==0.2.2
hypothesis==6.27.3
idna==3.3
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
py-solc==3.2.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycryptodome==3.14.1
Pygments==2.11.2
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-decouple==3.6
python-dotenv==0.16.0
pythx==1.6.1
pywin32==303
PyYAML==5.4.1
requests==2.27.1
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
varint==1.0.2
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.27.0
websockets==9.1
wrapt==1.13.3
yarl==1.7.2